+++ lib/src/types.dart
@@ -1,0 +1,39 @@
+import 'package:flutter/widgets.dart';
+
+enum PreferOrientation { left, right, top, bottom }
+
+class DefTooltipType {
+  final String message;
+  final double height;
+  final EdgeInsetsGeometry? padding;
+  final EdgeInsetsGeometry? margin;
+  final Decoration? decoration;
+  final TextStyle? textStyle;
+  final Animation<double> animation;
+  final Offset target;
+  final double allOffset;
+  final PreferOrientation preferOri;
+  final OverlayEntry entry;
+  final Size targetSize;
+  final Function customDismiss;
+  final Color? triangleColor;
+  final bool ignorePointer;
+  DefTooltipType(
+      {Key? key,
+      required this.message,
+      required this.height,
+      bool? ignorePointer,
+      this.triangleColor,
+      this.padding,
+      this.margin,
+      this.decoration,
+      this.textStyle,
+      required this.animation,
+      required this.target,
+      required this.allOffset,
+      required this.preferOri,
+      required this.entry,
+      required this.targetSize,
+      required this.customDismiss})
+      : ignorePointer = ignorePointer ?? false;
+}
