--- lib/types.dart
@@ -1,1 +1,0 @@
-enum PreferOrientation { left, right, top, bottom }
